+++ src/main/java/com/mattfuncional/config/SecurityConfig.java
@@ -52,7 +52,7 @@ public class SecurityConfig {
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
         http
                 .authorizeHttpRequests(authorize -> authorize
-                        .requestMatchers("/login", "/css/**", "/js/**", "/img/**", "/", "/error", "/demo", "/publica", "/planes", "/public/**").permitAll()
+                        .requestMatchers("/login", "/css/**", "/js/**", "/img/**", "/media/promociones/**", "/", "/error", "/demo", "/publica", "/planes", "/public/**").permitAll()
                         .requestMatchers("/rutinas/hoja/**").permitAll()
                         .requestMatchers("/sala/**").permitAll()
                         .requestMatchers("/profesor/usuarios-sistema/**", "/profesor/pagina-publica/**").hasAnyRole("ADMIN", "DEVELOPER")